move fixtures to search folder
